+++ ndk-context/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add `release_android_context()` function to remove `AndroidContext` when activity is destroyed. (#263)
+
 # 0.1.0 (2022-02-14)
 
 - Initial release! 🎉